+++ src/DDLQueryBuilder.php
@@ -83,6 +83,19 @@ final class DDLQueryBuilder extends AbstractDDLQueryBuilder
         return 'DROP INDEX ' . $this->quoter->quoteTableName($name);
     }
 
+    /**
+     * @throws NotSupportedException Oracle doesn't support "IF EXISTS" option on drop table.
+     */
+    public function dropTable(string $table, bool $ifExists = false, bool $cascade = false): string
+    {
+        if ($ifExists) {
+            throw new NotSupportedException('Oracle doesn\'t support "IF EXISTS" option on drop table.');
+        }
+        return 'DROP TABLE '
+            . $this->quoter->quoteTableName($table)
+            . ($cascade ? ' CASCADE CONSTRAINTS' : '');
+    }
+
     public function renameTable(string $oldName, string $newName): string
     {
         return 'ALTER TABLE ' . $this->quoter->quoteTableName($oldName) . ' RENAME TO ' .